+++ CHANGELOG.md
@@ -1,9 +1,21 @@
 CHANGELOG
 =========
 
+<<<<<<< HEAD
 #### 2.0.1 (2017-11-27)
 
 * Adds support for array type in request filters.
+=======
+#### 2.1.0 (2017-11-29)
+
+* Fixed bug undefined constant in class method `Oka\PaginationBundle\Service\PagingationManager::paginate()`.
+* Added the ability of filtering query with sql expresion in URL query filters.
+* Added `oka_pagination.query_builder_manipulator` service ID.
+* Added `oka_pagination.query_expr_converters` configuration values.
+* Added `Oka\PaginationBundle\Converter\LikeQueryExprConverter` class in query expression map converter.
+* Added `Oka\PaginationBundle\Converter\NotLikeQueryExprConverter` class in query expression map converter.
+* Updated documentation.
+>>>>>>> 2.0
 
 #### 2.0.0 (2017-11-25)